Extracted PersistanceHelper as a module
