+++ app/build.gradle.kts
@@ -109,6 +109,7 @@ dependencies {
     implementation("org.kodein.di:kodein-di-jvm:7.6.0")
     implementation("com.squareup.okhttp3:okhttp:4.9.1")
 
+    androidTestImplementation(project(":platform:testing-persistance"))
     androidTestImplementation(project(":platform:testing-rest"))
 
     androidTestImplementation("junit:junit:4.13.2")
