+++ platform/testing-persistance/src/main/java/io/dotanuki/testing/persistance/LocalStorageExtensions.kt
@@ -1,0 +1,9 @@
+package io.dotanuki.testing.persistance
+
+import io.dotanuki.norris.persistance.LocalStorage
+import org.kodein.di.DIAware
+import org.kodein.di.direct
+import org.kodein.di.instance
+
+fun DIAware.localStorage(): LocalStorage =
+    di.direct.instance()
