+++ platform/testing-persistance/src/main/java/io/dotanuki/testing/persistance/PersistanceHelper.kt
@@ -1,7 +1,8 @@
-package io.dotanuki.demos.norris
+package io.dotanuki.testing.persistance
 
 import androidx.test.platform.app.InstrumentationRegistry
 import io.dotanuki.norris.persistance.LocalStorage
+import kotlinx.coroutines.runBlocking
 import org.kodein.di.DIAware
 import org.kodein.di.direct
 import org.kodein.di.instance
@@ -15,6 +16,8 @@ object PersistanceHelper {
     fun registerNewSearch(term: String) {
         retrieveStorage().registerNewSearch(term)
     }
+
+    fun savedSearches(): List<String> = runBlocking { retrieveStorage().lastSearches() }
 
     private fun retrieveStorage(): LocalStorage {
         val app = InstrumentationRegistry.getInstrumentation().targetContext.applicationContext
